requirements.txt
feedparser==6.0.11
requests==2.32.3
google-generativeai==0.8.3
tweepy==4.14.0
PyYAML==6.0.2
python-dateutil==2.9.0
